+++ guillotina/static/gmi/precache-manifest.c09bcd481c96ebe3bf5a8293084b3a3e.js
@@ -1,23 +1,23 @@
 self.__precacheManifest = (self.__precacheManifest || []).concat([
   {
-    "revision": "5e51d5f56a5a5495c8d1248773f0d08f",
+    "revision": "76704c74dad07f50e5c73855f2ba5a7a",
     "url": "/+manage/index.html"
   },
   {
-    "revision": "e7d2b8822c469568848e",
+    "revision": "3641e795dccda559ada0",
     "url": "/+manage/static/css/main.fcbb1673.chunk.css"
   },
   {
-    "revision": "96bfac66e545962fdcb2",
-    "url": "/+manage/static/js/2.33f3bbfb.chunk.js"
+    "revision": "720d674a7fb82499e617",
+    "url": "/+manage/static/js/2.5f32d04e.chunk.js"
   },
   {
     "revision": "d705cb622423d72c5defbf368ca70dcc",
-    "url": "/+manage/static/js/2.33f3bbfb.chunk.js.LICENSE"
+    "url": "/+manage/static/js/2.5f32d04e.chunk.js.LICENSE"
   },
   {
-    "revision": "e7d2b8822c469568848e",
-    "url": "/+manage/static/js/main.7b009301.chunk.js"
+    "revision": "3641e795dccda559ada0",
+    "url": "/+manage/static/js/main.2bfc59fe.chunk.js"
   },
   {
     "revision": "da4ef6d8a4e37fe5c0fe",