+++ guillotina/tests/test_swagger.py
@@ -1,7 +1,9 @@
+from guillotina.api.service import _safe_int_or_float_cast
 from openapi_spec_validator import validate_v3_spec
 
 import os
 import pytest
+import unittest
 
 
 pytestmark = pytest.mark.asyncio
@@ -65,3 +67,50 @@ async def test_validate_arrays_in_query_params(container_requester):
         )
         assert status == 412
         assert resp["validator"] == "maxItems"
+
+
+@pytest.mark.app_settings(SWAGGER_SETTINGS)
+async def test_validate_numbers_and_integers_in_query_params(container_requester):
+    async with container_requester as requester:
+        _, status = await requester("GET", "@queryParamsValidation?oranges=5")
+        assert status == 200
+
+        _, status = await requester("GET", "@queryParamsValidation?kilograms=60.3")
+        assert status == 200
+
+        resp, status = await requester("GET", "@queryParamsValidation?oranges=0")
+        assert status == 412
+        assert resp["validator"] == "minimum"
+
+        resp, status = await requester("GET", "@queryParamsValidation?kilograms=3.2")
+        assert status == 412
+        assert resp["validator"] == "minimum"
+
+
+class Test_SafeIntOrFloatCast(unittest.TestCase):
+    def _makeOne(self, value):
+        return _safe_int_or_float_cast(value)
+
+    def test_with_integers(self):
+        value = self._makeOne(2)
+        self.assertIsInstance(value, int)
+        self.assertEqual(2, value)
+
+    def test_with_integers_from_string(self):
+        value = self._makeOne("2")
+        self.assertIsInstance(value, int)
+        self.assertEqual(2, value)
+
+    def test_with_floats(self):
+        value = self._makeOne(3.3)
+        self.assertIsInstance(value, float)
+        self.assertEqual(3.3, value)
+
+    def test_with_floats_from_string(self):
+        value = self._makeOne("3.0")
+        self.assertIsInstance(value, float)
+        self.assertEqual(3.0, value)
+
+    def test_returns_input_if_cannot_cast(self):
+        for foo in ([], None, {}, set(), ""):
+            self.assertIs(foo, self._makeOne(foo))